tests/wpt/web-platform-tests/content-security-policy/style-src/style-src-injected-inline-style-allowed.html

<!DOCTYPE HTML>
<html>
<head>
    <meta http-equiv="Content-Security-Policy" content="script-src 'self' 'unsafe-inline'; style-src 'unsafe-inline'">
    <script src='/resources/testharness.js'></script>
    <script src='/resources/testharnessreport.js'></script>

    <script>
        var t = async_test("Injected inline style should load with 'unsafe-inline'");
        document.addEventListener("securitypolicyviolation", t.unreached_func("Should not trigger a security policy violation"));
    </script>

</head>
<body>
    <div id='log'></div>

    <div id="content">Lorem ipsum</div>

    <script src='/content-security-policy/style-src/resources/style-src-inject-style.js'></script>

    <script>
      t.step(function() {
        onload = t.step_func_done(function(e) {
          var contentEl = document.getElementById("content");
          var marginLeftVal = getComputedStyle(contentEl).getPropertyValue('margin-left');
          assert_true(marginLeftVal == "2px");
          var marginRightVal = getComputedStyle(contentEl).getPropertyValue('margin-right');
          assert_true(marginRightVal == "2px");
        });
      });
    </script>

</body>
</html>
